Better Node testability
 - inject serializer
 - fixed gzip ser
 - more Node tests
 - renamed test assembly to .UTest

+++ kafka-sharp/kafka-sharp/Public/Configuration.cs
@@ -70,7 +70,7 @@ namespace Kafka.Public
         /// <summary>
         /// Acknowledgements required.
         /// </summary>
-        public int RequiredAcks = 1;
+        public short RequiredAcks = 1;
 
         /// <summary>
         /// Kafka server side timeout for requests.
